Make fILE IS Compiling - Type mismatch error is fixed

+++ Erebor/nqp_printer.c
@@ -5,42 +5,42 @@
 #include "nqp_thread.h"
 #include "nqp_thread_sched.h"
 
-void printer( void *arg )
+void printer(void *arg)
 {
-    assert( arg != NULL );
+    assert(arg != NULL);
 
     uint64_t counter = 0;
     char letter = '\0';
 
-    if ( arg != NULL )
+    if (arg != NULL)
     {
         letter = *(char *)arg;
-        while ( counter < 200 )
+        while (counter < 200)
         {
-            printf("%c %lu\n", letter, counter );
+            printf("%c %llu\n", letter, (unsigned long long)counter);
             counter++;
 
-            if ( counter % 50 == 0 )
+            if (counter % 50 == 0)
             {
-                nqp_yield( );
+                nqp_yield();
             }
         }
     }
 
-    nqp_exit( );
+    nqp_exit();
 }
 
-int main( void )
+int main(void)
 {
     char *letters = "ab";
-    
-    nqp_thread_t *thread_a = nqp_thread_create( printer, &letters[0] );
-    nqp_thread_t *thread_b = nqp_thread_create( printer, &letters[1] );
+
+    nqp_thread_t *thread_a = nqp_thread_create(printer, &letters[0]);
+    nqp_thread_t *thread_b = nqp_thread_create(printer, &letters[1]);
 
     nqp_sched_start();
 
-    nqp_thread_join( thread_a );
-    nqp_thread_join( thread_b );
+    nqp_thread_join(thread_a);
+    nqp_thread_join(thread_b);
 
     return EXIT_SUCCESS;
 }
